Tests/LibWeb/Text/expected/CookieChangeEvent-cached-attributes.txt

changed same object: true
deleted same object: true
changed frozen: true
deleted frozen: true
changed name: session
changed value: fresh
deleted name: old-session
deleted value: undefined
